Implement viewing and adding to cupboard in frontend, and necessary API endpoints
Adds API endpoints to get a user's cupboard, add to it and remove from it, and to get and delete ingredients. User's cupboard is shown in frontend with collapsible accordions and an input is added at the top of the page to add an ingredient to the cupboard. Also, access token is now created on login with str(user.id) since the identity must be a string.