Implement viewing and adding recipes on frontend and backend
Adds API endpoints to get, create, edit and delete recipes. On frontend, adds cards for each recipe showing what ingredients the user is missing and a button which opens a modal to add a recipe. Removes the image column from the recipe table, to be implemented at a later date.