Handle JWT expirations properly and improve error handling on API endpoints
When a JWT expires in the frontend, the user is logged out and redirected to the login page. To improve error handling of database commits, adds a safe_commit function to utils.py which every API endpoint which alters the db calls - this attempts to commit and rollsback on failure.